+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Refactor logic for identifying granule spatial representation. (Issue-159)
 * End processing if UMM-C metadata can't be retrieved and/or don't contain a
   `GranuleSpatialRepresentation` element.
+* Add support for bounding rectangles. (Issue-157)
+
 
 ## v1.6.0